requirements.txt
torch
torchvision
transformers~=4.49.0
datasets~=3.4.0
flask~=3.1.0
pillow~=11.1.0
timm
peft
einops
scikit-learn
pandas~=2.2.3